+++ nextgisweb/lookup_table/nodepkg/editor/index.ts
@@ -1,0 +1,3 @@
+/** @entrypoint */
+export { EditorStore as store } from "./EditorStore";
+export { EditorWidget as widget } from "./EditorWidget";